embedder/requirements.txt
grpcio==1.41.0
grpcio-tools==1.41.0
sentence-transformers==2.2.2